Change style for trainer persona
When a user logs in a trainer there should be a visual change that
suggests that user is now a trainer.

* app/assets/stylesheets/application.css.scss
require_tree . includes all css files in the given directory.  Since we
want trainer to be its own tree and not be part of application.css,
changed it the code to explicitly require files to be included

* app/assets/stylesheets/trainer.css.scss
This is a copy of application.css with different colors for navbar (for
now) that distinguishes regular pages from trainer pages

* app/controllers/trainer_controller.rb
Landing (stub) page for trainer dashboard

* app/views/layouts/trainer.html.erb
Layout for trainer (similar to application for now, but should change
in future to look different for trainer users)

* app/views/trainer/dashboard.html.erb
Stub for trainer dashboard view

* config/environments/development.rb
* config/environments/production.rb
Explicitly add trainer.css to precompile array since it is not part of
application.css

* config/routes.rb
Add route for GET trainer

NOTE: Perhaps it is better to flip and have application.rb to cater for
trainer instead so the changes are minimal.  But there will be a
similar requirement to change the root css for mom user as well.